refactor(workflow_client): make api_constants module private
Since the module is used by the client itself, do not make it public for
imports.

This also removes it as a `Submodules` entry in the API Reference page
in sphinx docs.
