+++ tests/workflow_client/test_response_error.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, Any
 import pytest
 
 from exb_sdk.workflow_client import Client
-from exb_sdk.workflow_client.api_constants import DocumentState
+from exb_sdk.workflow_client._api_constants import DocumentState  # noqa: PLC2701
 from exb_sdk.workflow_client.exceptions import HttpError
 from tests.workflow_client.conftest import Context, response_state, response_upload
